Stage 20: assignment 2
Expl pgm to which creates a linked list which stores numbers from 1 to 100 in each node.
Then fork the process such that the parent will print the odd number (1st, 3rd...) and the child will print the even numbers (0th, 2nd ..)

You will notice that the numbers printed will not be in order.

Batch file is updated to run this process as the init process